# perf: refactor interpreter internals