--- a/bins/revme/src/statetest/runner.rs
+++ b/bins/revme/src/statetest/runner.rs
@@ -110,7 +110,7 @@ pub fn execute_test_suit(
     let json_reader = std::fs::read(path).unwrap();
     let suit: TestSuit = serde_json::from_reader(&*json_reader)?;
 
-    let map_caller_keys: HashMap<_, _> = vec![
+    let map_caller_keys: HashMap<_, _> = [
         (
             B256(hex!(
                 "45a915e4d060149eb4365960e6a7a45f334393093061116b197e3240065ff2d8"