--- a/crates/interpreter/src/gas/calc.rs
+++ b/crates/interpreter/src/gas/calc.rs
@@ -1,13 +1,13 @@
 use super::constants::*;
 use crate::inner_models::SelfDestructResult;
-use crate::primitives::{Bytes, SpecId, SpecId::*, B160, U256};
+use crate::primitives::{Bytes, Spec, SpecId::*, B160, U256};
 use alloc::vec::Vec;
 
 #[allow(clippy::collapsible_else_if)]
-pub fn sstore_refund(original: U256, current: U256, new: U256, spec: SpecId) -> i64 {
-    if SpecId::enabled(spec, ISTANBUL) {
+pub fn sstore_refund<SPEC: Spec>(original: U256, current: U256, new: U256) -> i64 {
+    if SPEC::enabled(ISTANBUL) {
         // EIP-3529: Reduction in refunds
-        let sstore_clears_schedule = if SpecId::enabled(spec, LONDON) {
+        let sstore_clears_schedule = if SPEC::enabled(LONDON) {
             (SSTORE_RESET - COLD_SLOAD_COST + ACCESS_LIST_STORAGE_KEY) as i64
         } else {
             REFUND_SSTORE_CLEARS
@@ -29,10 +29,10 @@ pub fn sstore_refund(original: U256, current: U256, new: U256, spec: SpecId) -> 
                 }
 
                 if original == new {
-                    let (gas_sstore_reset, gas_sload) = if SpecId::enabled(spec, BERLIN) {
+                    let (gas_sstore_reset, gas_sload) = if SPEC::enabled(BERLIN) {
                         (SSTORE_RESET - COLD_SLOAD_COST, WARM_STORAGE_READ_COST)
                     } else {
-                        (SSTORE_RESET, sload_cost(false, spec))
+                        (SSTORE_RESET, sload_cost::<SPEC>(false))
                     };
                     if original == U256::ZERO {
                         refund += (SSTORE_SET - gas_sload) as i64;
@@ -86,12 +86,12 @@ fn log2floor(value: U256) -> u64 {
 }
 
 #[inline]
-pub fn exp_cost(power: U256, spec: SpecId) -> Option<u64> {
+pub fn exp_cost<SPEC: Spec>(power: U256) -> Option<u64> {
     if power == U256::ZERO {
         Some(EXP)
     } else {
         // EIP-160: EXP cost increase
-        let gas_byte = U256::from(if SpecId::enabled(spec, SPURIOUS_DRAGON) {
+        let gas_byte = U256::from(if SPEC::enabled(SPURIOUS_DRAGON) {
             50u64
         } else {
             10
@@ -111,17 +111,17 @@ pub fn verylowcopy_cost(len: u64) -> Option<u64> {
 }
 
 #[inline]
-pub fn extcodecopy_cost(len: u64, is_cold: bool, spec: SpecId) -> Option<u64> {
+pub fn extcodecopy_cost<SPEC: Spec>(len: u64, is_cold: bool) -> Option<u64> {
     let wordd = len / 32;
     let wordr = len % 32;
 
-    let base_gas: u64 = if SpecId::enabled(spec, BERLIN) {
+    let base_gas: u64 = if SPEC::enabled(BERLIN) {
         if is_cold {
             COLD_ACCOUNT_ACCESS_COST
         } else {
             WARM_STORAGE_READ_COST
         }
-    } else if SpecId::enabled(spec, TANGERINE) {
+    } else if SPEC::enabled(TANGERINE) {
         700
     } else {
         20
@@ -129,14 +129,14 @@ pub fn extcodecopy_cost(len: u64, is_cold: bool, spec: SpecId) -> Option<u64> {
     base_gas.checked_add(COPY.checked_mul(if wordr == 0 { wordd } else { wordd + 1 })?)
 }
 
-pub fn account_access_gas(is_cold: bool, spec: SpecId) -> u64 {
-    if SpecId::enabled(spec, BERLIN) {
+pub fn account_access_gas<SPEC: Spec>(is_cold: bool) -> u64 {
+    if SPEC::enabled(BERLIN) {
         if is_cold {
             COLD_ACCOUNT_ACCESS_COST
         } else {
             WARM_STORAGE_READ_COST
         }
-    } else if SpecId::enabled(spec, ISTANBUL) {
+    } else if SPEC::enabled(ISTANBUL) {
         700
     } else {
         20
@@ -167,17 +167,17 @@ pub fn initcode_cost(len: u64) -> u64 {
 }
 
 #[inline]
-pub fn sload_cost(is_cold: bool, spec: SpecId) -> u64 {
-    if SpecId::enabled(spec, BERLIN) {
+pub fn sload_cost<SPEC: Spec>(is_cold: bool) -> u64 {
+    if SPEC::enabled(BERLIN) {
         if is_cold {
             COLD_SLOAD_COST
         } else {
             WARM_STORAGE_READ_COST
         }
-    } else if SpecId::enabled(spec, ISTANBUL) {
+    } else if SPEC::enabled(ISTANBUL) {
         // EIP-1884: Repricing for trie-size-dependent opcodes
         800
-    } else if SpecId::enabled(spec, TANGERINE) {
+    } else if SPEC::enabled(TANGERINE) {
         // EIP-150: Gas cost changes for IO-heavy operations
         200
     } else {
@@ -186,24 +186,23 @@ pub fn sload_cost(is_cold: bool, spec: SpecId) -> u64 {
 }
 
 #[allow(clippy::collapsible_else_if)]
-pub fn sstore_cost(
+pub fn sstore_cost<SPEC: Spec>(
     original: U256,
     current: U256,
     new: U256,
     gas: u64,
     is_cold: bool,
-    spec: SpecId,
 ) -> Option<u64> {
     // TODO untangle this mess and make it more elegant
-    let (gas_sload, gas_sstore_reset) = if SpecId::enabled(spec, BERLIN) {
+    let (gas_sload, gas_sstore_reset) = if SPEC::enabled(BERLIN) {
         (WARM_STORAGE_READ_COST, SSTORE_RESET - COLD_SLOAD_COST)
     } else {
-        (sload_cost(is_cold, spec), SSTORE_RESET)
+        (sload_cost::<SPEC>(is_cold), SSTORE_RESET)
     };
 
     // https://eips.ethereum.org/EIPS/eip-2200
     // It’s a combined version of EIP-1283 and EIP-1706
-    let gas_cost = if SpecId::enabled(spec, ISTANBUL) {
+    let gas_cost = if SPEC::enabled(ISTANBUL) {
         // EIP-1706
         if gas <= CALL_STIPEND {
             return None;
@@ -231,59 +230,54 @@ pub fn sstore_cost(
         }
     };
     // In EIP-2929 we charge extra if the slot has not been used yet in this transaction
-    if SpecId::enabled(spec, BERLIN) && is_cold {
+    if SPEC::enabled(BERLIN) && is_cold {
         Some(gas_cost + COLD_SLOAD_COST)
     } else {
         Some(gas_cost)
     }
 }
 
-pub fn selfdestruct_cost(res: SelfDestructResult, spec: SpecId) -> u64 {
+pub fn selfdestruct_cost<SPEC: Spec>(res: SelfDestructResult) -> u64 {
     // EIP-161: State trie clearing (invariant-preserving alternative)
-    let should_charge_topup = if SpecId::enabled(spec, SPURIOUS_DRAGON) {
+    let should_charge_topup = if SPEC::enabled(SPURIOUS_DRAGON) {
         res.had_value && !res.target_exists
     } else {
         !res.target_exists
     };
 
     // EIP-150: Gas cost changes for IO-heavy operations
-    let selfdestruct_gas_topup = if SpecId::enabled(spec, TANGERINE) && should_charge_topup {
+    let selfdestruct_gas_topup = if SPEC::enabled(TANGERINE) && should_charge_topup {
         25000
     } else {
         0
     };
 
     // EIP-150: Gas cost changes for IO-heavy operations
-    let selfdestruct_gas = if SpecId::enabled(spec, TANGERINE) {
-        5000
-    } else {
-        0
-    };
+    let selfdestruct_gas = if SPEC::enabled(TANGERINE) { 5000 } else { 0 };
 
     let mut gas = selfdestruct_gas + selfdestruct_gas_topup;
-    if SpecId::enabled(spec, BERLIN) && res.is_cold {
+    if SPEC::enabled(BERLIN) && res.is_cold {
         gas += COLD_ACCOUNT_ACCESS_COST
     }
     gas
 }
 
-pub fn call_cost(
+pub fn call_cost<SPEC: Spec>(
     value: U256,
     is_new: bool,
     is_cold: bool,
     is_call_or_callcode: bool,
     is_call_or_staticcall: bool,
-    spec: SpecId,
 ) -> u64 {
     let transfers_value = value != U256::default();
 
-    let call_gas = if SpecId::enabled(spec, BERLIN) {
+    let call_gas = if SPEC::enabled(BERLIN) {
         if is_cold {
             COLD_ACCOUNT_ACCESS_COST
         } else {
             WARM_STORAGE_READ_COST
         }
-    } else if SpecId::enabled(spec, TANGERINE) {
+    } else if SPEC::enabled(TANGERINE) {
         // EIP-150: Gas cost changes for IO-heavy operations
         700
     } else {
@@ -292,12 +286,12 @@ pub fn call_cost(
 
     call_gas
         + xfer_cost(is_call_or_callcode, transfers_value)
-        + new_cost(is_call_or_staticcall, is_new, transfers_value, spec)
+        + new_cost::<SPEC>(is_call_or_staticcall, is_new, transfers_value)
 }
 
 #[inline]
-pub fn hot_cold_cost(is_cold: bool, regular_value: u64, spec: SpecId) -> u64 {
-    if SpecId::enabled(spec, BERLIN) {
+pub fn hot_cold_cost<SPEC: Spec>(is_cold: bool, regular_value: u64) -> u64 {
+    if SPEC::enabled(BERLIN) {
         if is_cold {
             COLD_ACCOUNT_ACCESS_COST
         } else {
@@ -318,10 +312,10 @@ fn xfer_cost(is_call_or_callcode: bool, transfers_value: bool) -> u64 {
 }
 
 #[inline]
-fn new_cost(is_call_or_staticcall: bool, is_new: bool, transfers_value: bool, spec: SpecId) -> u64 {
+fn new_cost<SPEC: Spec>(is_call_or_staticcall: bool, is_new: bool, transfers_value: bool) -> u64 {
     if is_call_or_staticcall {
         // EIP-161: State trie clearing (invariant-preserving alternative)
-        if SpecId::enabled(spec, SPURIOUS_DRAGON) {
+        if SPEC::enabled(SPURIOUS_DRAGON) {
             if transfers_value && is_new {
                 NEWACCOUNT
             } else {
@@ -347,11 +341,10 @@ pub fn memory_gas(a: usize) -> u64 {
 
 /// Initial gas that is deducted for transaction to be included.
 /// Initial gas contains initial stipend gas, gas for access list and input data.
-pub fn initial_tx_gas(
+pub fn initial_tx_gas<SPEC: Spec>(
     input: &Bytes,
     is_create: bool,
     access_list: &[(B160, Vec<U256>)],
-    spec: SpecId,
 ) -> u64 {
     let mut initial_gas = 0;
     let zero_data_len = input.iter().filter(|v| **v == 0).count() as u64;
@@ -360,15 +353,10 @@ pub fn initial_tx_gas(
     // initdate stipend
     initial_gas += zero_data_len * TRANSACTION_ZERO_DATA;
     // EIP-2028: Transaction data gas cost reduction
-    initial_gas += non_zero_data_len
-        * if SpecId::enabled(spec, ISTANBUL) {
-            16
-        } else {
-            68
-        };
+    initial_gas += non_zero_data_len * if SPEC::enabled(ISTANBUL) { 16 } else { 68 };
 
     // get number of access list account and storages.
-    if SpecId::enabled(spec, BERLIN) {
+    if SPEC::enabled(BERLIN) {
         let accessed_slots = access_list
             .iter()
             .fold(0, |slot_count, (_, slots)| slot_count + slots.len() as u64);
@@ -378,7 +366,7 @@ pub fn initial_tx_gas(
 
     // base stipend
     initial_gas += if is_create {
-        if SpecId::enabled(spec, HOMESTEAD) {
+        if SPEC::enabled(HOMESTEAD) {
             // EIP-2: Homestead Hard-fork Changes
             53000
         } else {
@@ -390,7 +378,7 @@ pub fn initial_tx_gas(
 
     // EIP-3860: Limit and meter initcode
     // Initcode stipend for bytecode analysis
-    if SpecId::enabled(spec, SHANGHAI) && is_create {
+    if SPEC::enabled(SHANGHAI) && is_create {
         initial_gas += initcode_cost(input.len() as u64)
     }