--- a/crates/interpreter/src/host/mod.rs
+++ b/crates/interpreter/src/host/mod.rs
@@ -1,12 +1,12 @@
-mod dummy_host;
+mod dummy;
 
 use crate::primitives::Bytecode;
 use crate::{
     primitives::{Bytes, Env, B160, B256, U256},
     CallInputs, CreateInputs, Gas, InstructionResult, Interpreter, SelfDestructResult,
 };
-pub use alloc::vec::Vec;
-pub use dummy_host::DummyHost;
+use alloc::vec::Vec;
+pub use dummy::DummyHost;
 
 /// EVM context host.
 pub trait Host {