… Cargo.lock
… Cargo.toml
… bins/revm-test/Cargo.toml
… bins/revm-test/src/bin/analysis.rs
… bins/revm-test/src/bin/transfer.rs
… bins/revme/Cargo.toml
… crates/interpreter/Cargo.toml
… crates/interpreter/src/gas/mod.rs
… crates/interpreter/src/host/dummy.rs
… crates/interpreter/src/instructions.rs
… crates/interpreter/src/instructions/arithmetic.rs
… crates/interpreter/src/instructions/bitwise.rs
… crates/interpreter/src/instructions/control.rs
… crates/interpreter/src/instructions/host.rs
… crates/interpreter/src/instructions/host_env.rs
… crates/interpreter/src/instructions/i256.rs
… crates/interpreter/src/instructions/macros.rs
… crates/interpreter/src/instructions/memory.rs
… crates/interpreter/src/instructions/mod.rs
… crates/interpreter/src/instructions/opcode.rs
… crates/interpreter/src/instructions/stack.rs
… crates/interpreter/src/instructions/system.rs
… crates/interpreter/src/interpreter/analysis.rs
… crates/interpreter/src/interpreter/contract.rs
… crates/interpreter/src/interpreter/memory.rs
… crates/interpreter/src/interpreter/mod.rs
… crates/interpreter/src/interpreter/stack.rs
… crates/interpreter/src/lib.rs
… crates/precompile/src/lib.rs
… crates/precompile/src/modexp.rs