This PR focuses on refactoring some parts of the interpreter which lead to, although minor, performance gains.

Closes #310, not fixes, as the current `match` performs better on debug builds, and performs the same in release builds. This is not shown in the results.

# Results

## Cachegrind

[cachegrind.tar.gz](https://github.com/bluealloy/revm/files/12210142/cachegrind.tar.gz)

|                      Commit                     | Snailtracer (I) |
|:-----------------------------------------------:|:---------------:|
| 968fe98fcb590e5d67d690ce3c6a370834dc0541 (main) |   515,151,261   |
|     2a06916ef30e93b9b23907624b8e04ee63994693    |   489,886,182   |
| Improvement                                     | -4.9%           |

## Benches

~~Not entirely sure why the non-generic version performs better, since I would expect the checks to be optimized out at compile time on the generic version.~~

~~I have included the patch which can be used to change between the two versions.~~

Time benches shouldn't be relied upon for these type of small changes, and these ones are also outdated.

### Main

36de35b25d5941ac257e1877696e894223445c30

```julia
snailtracer/transact/analysed                                                                           
                        time:   [46.154 ms 46.666 ms 47.183 ms]
snailtracer/eval        time:   [40.100 ms 40.293 ms 40.487 ms]                            


snailtracer/transact/analysed                                                                           
                        time:   [45.875 ms 46.356 ms 47.026 ms]
snailtracer/eval        time:   [40.874 ms 41.289 ms 41.946 ms]                            


snailtracer/transact/analysed                                                                           
                        time:   [46.696 ms 46.993 ms 47.451 ms]
snailtracer/eval        time:   [40.366 ms 40.707 ms 41.054 ms]                            
```

### Generic (patch)

3658fa7a16e2dc3ea7b125bd836380f4a3ed678f

<details><summary><strong>Patch</strong></summary>
<p>

```diff
diff --git a/crates/interpreter/src/gas/calc.rs b/crates/interpreter/src/gas/calc.rs
index fa93b6a..dbd05be 100644
--- a/crates/interpreter/src/gas/calc.rs
+++ b/crates/interpreter/src/gas/calc.rs
@@ -1,13 +1,13 @@
 use super::constants::*;
 use crate::inner_models::SelfDestructResult;
 use alloc::vec::Vec;
-use primitives::{Bytes, SpecId, SpecId::*, B160, U256};
+use primitives::{Bytes, Spec, SpecId::*, B160, U256};
 
 #[allow(clippy::collapsible_else_if)]
-pub fn sstore_refund(original: U256, current: U256, new: U256, spec: SpecId) -> i64 {
-    if SpecId::enabled(spec, ISTANBUL) {
+pub fn sstore_refund<SPEC: Spec>(original: U256, current: U256, new: U256) -> i64 {
+    if SPEC::enabled(ISTANBUL) {
         // EIP-3529: Reduction in refunds
-        let sstore_clears_schedule = if SpecId::enabled(spec, LONDON) {
+        let sstore_clears_schedule = if SPEC::enabled(LONDON) {
             (SSTORE_RESET - COLD_SLOAD_COST + ACCESS_LIST_STORAGE_KEY) as i64
         } else {
             REFUND_SSTORE_CLEARS
@@ -29,10 +29,10 @@ pub fn sstore_refund(original: U256, current: U256, new: U256, spec: SpecId) ->
                 }
 
                 if original == new {
-                    let (gas_sstore_reset, gas_sload) = if SpecId::enabled(spec, BERLIN) {
+                    let (gas_sstore_reset, gas_sload) = if SPEC::enabled(BERLIN) {
                         (SSTORE_RESET - COLD_SLOAD_COST, WARM_STORAGE_READ_COST)
                     } else {
-                        (SSTORE_RESET, sload_cost(false, spec))
+                        (SSTORE_RESET, sload_cost::<SPEC>(false))
                     };
                     if original == U256::ZERO {
                         refund += (SSTORE_SET - gas_sload) as i64;
@@ -83,12 +83,12 @@ fn log2floor(value: U256) -> u64 {
     l
 }
 
-pub fn exp_cost(power: U256, spec: SpecId) -> Option<u64> {
+pub fn exp_cost<SPEC: Spec>(power: U256) -> Option<u64> {
     if power == U256::ZERO {
         Some(EXP)
     } else {
         // EIP-160: EXP cost increase
-        let gas_byte = U256::from(if SpecId::enabled(spec, SPURIOUS_DRAGON) {
+        let gas_byte = U256::from(if SPEC::enabled(SPURIOUS_DRAGON) {
             50u64
         } else {
             10
@@ -106,17 +106,17 @@ pub fn verylowcopy_cost(len: u64) -> Option<u64> {
     VERYLOW.checked_add(COPY.checked_mul(if wordr == 0 { wordd } else { wordd + 1 })?)
 }
 
-pub fn extcodecopy_cost(len: u64, is_cold: bool, spec: SpecId) -> Option<u64> {
+pub fn extcodecopy_cost<SPEC: Spec>(len: u64, is_cold: bool) -> Option<u64> {
     let wordd = len / 32;
     let wordr = len % 32;
 
-    let base_gas: u64 = if SpecId::enabled(spec, BERLIN) {
+    let base_gas: u64 = if SPEC::enabled(BERLIN) {
         if is_cold {
             COLD_ACCOUNT_ACCESS_COST
         } else {
             WARM_STORAGE_READ_COST
         }
-    } else if SpecId::enabled(spec, TANGERINE) {
+    } else if SPEC::enabled(TANGERINE) {
         700
     } else {
         20
@@ -124,14 +124,14 @@ pub fn extcodecopy_cost(len: u64, is_cold: bool, spec: SpecId) -> Option<u64> {
     base_gas.checked_add(COPY.checked_mul(if wordr == 0 { wordd } else { wordd + 1 })?)
 }
 
-pub fn account_access_gas(is_cold: bool, spec: SpecId) -> u64 {
-    if SpecId::enabled(spec, BERLIN) {
+pub fn account_access_gas<SPEC: Spec>(is_cold: bool) -> u64 {
+    if SPEC::enabled(BERLIN) {
         if is_cold {
             COLD_ACCOUNT_ACCESS_COST
         } else {
             WARM_STORAGE_READ_COST
         }
-    } else if SpecId::enabled(spec, ISTANBUL) {
+    } else if SPEC::enabled(ISTANBUL) {
         700
     } else {
         20
@@ -160,17 +160,17 @@ pub fn initcode_cost(len: u64) -> u64 {
     INITCODE_WORD_COST * if wordr == 0 { wordd } else { wordd + 1 }
 }
 
-pub fn sload_cost(is_cold: bool, spec: SpecId) -> u64 {
-    if SpecId::enabled(spec, BERLIN) {
+pub fn sload_cost<SPEC: Spec>(is_cold: bool) -> u64 {
+    if SPEC::enabled(BERLIN) {
         if is_cold {
             COLD_SLOAD_COST
         } else {
             WARM_STORAGE_READ_COST
         }
-    } else if SpecId::enabled(spec, ISTANBUL) {
+    } else if SPEC::enabled(ISTANBUL) {
         // EIP-1884: Repricing for trie-size-dependent opcodes
         800
-    } else if SpecId::enabled(spec, TANGERINE) {
+    } else if SPEC::enabled(TANGERINE) {
         // EIP-150: Gas cost changes for IO-heavy operations
         200
     } else {
@@ -179,24 +179,23 @@ pub fn sload_cost(is_cold: bool, spec: SpecId) -> u64 {
 }
 
 #[allow(clippy::collapsible_else_if)]
-pub fn sstore_cost(
+pub fn sstore_cost<SPEC: Spec>(
     original: U256,
     current: U256,
     new: U256,
     gas: u64,
     is_cold: bool,
-    spec: SpecId,
 ) -> Option<u64> {
     // TODO untangle this mess and make it more elegant
-    let (gas_sload, gas_sstore_reset) = if SpecId::enabled(spec, BERLIN) {
+    let (gas_sload, gas_sstore_reset) = if SPEC::enabled(BERLIN) {
         (WARM_STORAGE_READ_COST, SSTORE_RESET - COLD_SLOAD_COST)
     } else {
-        (sload_cost(is_cold, spec), SSTORE_RESET)
+        (sload_cost::<SPEC>(is_cold), SSTORE_RESET)
     };
 
     // https://eips.ethereum.org/EIPS/eip-2200
     // It’s a combined version of EIP-1283 and EIP-1706
-    let gas_cost = if SpecId::enabled(spec, ISTANBUL) {
+    let gas_cost = if SPEC::enabled(ISTANBUL) {
         // EIP-1706
         if gas <= CALL_STIPEND {
             return None;
@@ -224,58 +223,53 @@ pub fn sstore_cost(
         }
     };
     // In EIP-2929 we charge extra if the slot has not been used yet in this transaction
-    if SpecId::enabled(spec, BERLIN) && is_cold {
+    if SPEC::enabled(BERLIN) && is_cold {
         Some(gas_cost + COLD_SLOAD_COST)
     } else {
         Some(gas_cost)
     }
 }
 
-pub fn selfdestruct_cost(res: SelfDestructResult, spec: SpecId) -> u64 {
+pub fn selfdestruct_cost<SPEC: Spec>(res: SelfDestructResult) -> u64 {
     // EIP-161: State trie clearing (invariant-preserving alternative)
-    let should_charge_topup = if SpecId::enabled(spec, SPURIOUS_DRAGON) {
+    let should_charge_topup = if SPEC::enabled(SPURIOUS_DRAGON) {
         res.had_value && !res.target_exists
     } else {
         !res.target_exists
     };
 
-    let selfdestruct_gas_topup = if SpecId::enabled(spec, TANGERINE) && should_charge_topup {
+    let selfdestruct_gas_topup = if SPEC::enabled(TANGERINE) && should_charge_topup {
         //EIP-150: Gas cost changes for IO-heavy operations
         25000
     } else {
         0
     };
 
-    let selfdestruct_gas = if SpecId::enabled(spec, TANGERINE) {
-        5000
-    } else {
-        0
-    }; //EIP-150: Gas cost changes for IO-heavy operations
+    let selfdestruct_gas = if SPEC::enabled(TANGERINE) { 5000 } else { 0 }; //EIP-150: Gas cost changes for IO-heavy operations
 
     let mut gas = selfdestruct_gas + selfdestruct_gas_topup;
-    if SpecId::enabled(spec, BERLIN) && res.is_cold {
+    if SPEC::enabled(BERLIN) && res.is_cold {
         gas += COLD_ACCOUNT_ACCESS_COST
     }
     gas
 }
 
-pub fn call_cost(
+pub fn call_cost<SPEC: Spec>(
     value: U256,
     is_new: bool,
     is_cold: bool,
     is_call_or_callcode: bool,
     is_call_or_staticcall: bool,
-    spec: SpecId,
 ) -> u64 {
     let transfers_value = value != U256::default();
 
-    let call_gas = if SpecId::enabled(spec, BERLIN) {
+    let call_gas = if SPEC::enabled(BERLIN) {
         if is_cold {
             COLD_ACCOUNT_ACCESS_COST
         } else {
             WARM_STORAGE_READ_COST
         }
-    } else if SpecId::enabled(spec, TANGERINE) {
+    } else if SPEC::enabled(TANGERINE) {
         // EIP-150: Gas cost changes for IO-heavy operations
         700
     } else {
@@ -284,11 +278,11 @@ pub fn call_cost(
 
     call_gas
         + xfer_cost(is_call_or_callcode, transfers_value)
-        + new_cost(is_call_or_staticcall, is_new, transfers_value, spec)
+        + new_cost::<SPEC>(is_call_or_staticcall, is_new, transfers_value)
 }
 
-pub fn hot_cold_cost(is_cold: bool, regular_value: u64, spec: SpecId) -> u64 {
-    if SpecId::enabled(spec, BERLIN) {
+pub fn hot_cold_cost<SPEC: Spec>(is_cold: bool, regular_value: u64) -> u64 {
+    if SPEC::enabled(BERLIN) {
         if is_cold {
             COLD_ACCOUNT_ACCESS_COST
         } else {
@@ -307,10 +301,10 @@ fn xfer_cost(is_call_or_callcode: bool, transfers_value: bool) -> u64 {
     }
 }
 
-fn new_cost(is_call_or_staticcall: bool, is_new: bool, transfers_value: bool, spec: SpecId) -> u64 {
+fn new_cost<SPEC: Spec>(is_call_or_staticcall: bool, is_new: bool, transfers_value: bool) -> u64 {
     if is_call_or_staticcall {
         // EIP-161: State trie clearing (invariant-preserving alternative)
-        if SpecId::enabled(spec, SPURIOUS_DRAGON) {
+        if SPEC::enabled(SPURIOUS_DRAGON) {
             if transfers_value && is_new {
                 NEWACCOUNT
             } else {
@@ -335,11 +329,10 @@ pub fn memory_gas(a: usize) -> u64 {
 
 /// Initial gas that is deducted for transaction to be included.
 /// Initial gas contains initial stipend gas, gas for access list and input data.
-pub fn initial_tx_gas(
+pub fn initial_tx_gas<SPEC: Spec>(
     input: &Bytes,
     is_create: bool,
     access_list: &[(B160, Vec<U256>)],
-    spec: SpecId,
 ) -> u64 {
     let mut initial_gas = 0;
     let zero_data_len = input.iter().filter(|v| **v == 0).count() as u64;
@@ -348,15 +341,10 @@ pub fn initial_tx_gas(
     // initdate stipend
     initial_gas += zero_data_len * TRANSACTION_ZERO_DATA;
     // EIP-2028: Transaction data gas cost reduction
-    initial_gas += non_zero_data_len
-        * if SpecId::enabled(spec, ISTANBUL) {
-            16
-        } else {
-            68
-        };
+    initial_gas += non_zero_data_len * if SPEC::enabled(ISTANBUL) { 16 } else { 68 };
 
     // get number of access list account and storages.
-    if SpecId::enabled(spec, BERLIN) {
+    if SPEC::enabled(BERLIN) {
         let accessed_slots = access_list
             .iter()
             .fold(0, |slot_count, (_, slots)| slot_count + slots.len() as u64);
@@ -366,7 +354,7 @@ pub fn initial_tx_gas(
 
     // base stipend
     initial_gas += if is_create {
-        if SpecId::enabled(spec, HOMESTEAD) {
+        if SPEC::enabled(HOMESTEAD) {
             // EIP-2: Homestead Hard-fork Changes
             53000
         } else {
@@ -378,7 +366,7 @@ pub fn initial_tx_gas(
 
     // EIP-3860: Limit and meter initcode
     // Initcode stipend for bytecode analysis
-    if SpecId::enabled(spec, SHANGHAI) && is_create {
+    if SPEC::enabled(SHANGHAI) && is_create {
         initial_gas += initcode_cost(input.len() as u64)
     }
 
diff --git a/crates/interpreter/src/instructions/arithmetic.rs b/crates/interpreter/src/instructions/arithmetic.rs
index 8fa09ae..a93e1f9 100644
--- a/crates/interpreter/src/instructions/arithmetic.rs
+++ b/crates/interpreter/src/instructions/arithmetic.rs
@@ -1,25 +1,25 @@
 use super::i256::{i256_div, i256_mod};
 use super::prelude::*;
 
-pub(super) fn wrapped_add(interpreter: &mut Interpreter, _host: &mut dyn Host, _spec: SpecId) {
+pub(super) fn wrapped_add(interpreter: &mut Interpreter, _host: &mut dyn Host) {
     gas!(interpreter, gas::VERYLOW);
     pop_top!(interpreter, op1, op2);
     *op2 = op1.wrapping_add(*op2);
 }
 
-pub(super) fn wrapping_mul(interpreter: &mut Interpreter, _host: &mut dyn Host, _spec: SpecId) {
+pub(super) fn wrapping_mul(interpreter: &mut Interpreter, _host: &mut dyn Host) {
     gas!(interpreter, gas::LOW);
     pop_top!(interpreter, op1, op2);
     *op2 = op1.wrapping_mul(*op2);
 }
 
-pub(super) fn wrapping_sub(interpreter: &mut Interpreter, _host: &mut dyn Host, _spec: SpecId) {
+pub(super) fn wrapping_sub(interpreter: &mut Interpreter, _host: &mut dyn Host) {
     gas!(interpreter, gas::VERYLOW);
     pop_top!(interpreter, op1, op2);
     *op2 = op1.wrapping_sub(*op2);
 }
 
-pub(super) fn div(interpreter: &mut Interpreter, _host: &mut dyn Host, _spec: SpecId) {
+pub(super) fn div(interpreter: &mut Interpreter, _host: &mut dyn Host) {
     gas!(interpreter, gas::LOW);
     pop_top!(interpreter, op1, op2);
     if *op2 != U256::ZERO {
@@ -27,13 +27,13 @@ pub(super) fn div(interpreter: &mut Interpreter, _host: &mut dyn Host, _spec: Sp
     }
 }
 
-pub(super) fn sdiv(interpreter: &mut Interpreter, _host: &mut dyn Host, _spec: SpecId) {
+pub(super) fn sdiv(interpreter: &mut Interpreter, _host: &mut dyn Host) {
     gas!(interpreter, gas::LOW);
     pop_top!(interpreter, op1, op2);
     *op2 = i256_div(op1, *op2);
 }
 
-pub(super) fn rem(interpreter: &mut Interpreter, _host: &mut dyn Host, _spec: SpecId) {
+pub(super) fn rem(interpreter: &mut Interpreter, _host: &mut dyn Host) {
     gas!(interpreter, gas::LOW);
     pop_top!(interpreter, op1, op2);
     if *op2 != U256::ZERO {
@@ -41,7 +41,7 @@ pub(super) fn rem(interpreter: &mut Interpreter, _host: &mut dyn Host, _spec: Sp
     }
 }
 
-pub(super) fn smod(interpreter: &mut Interpreter, _host: &mut dyn Host, _spec: SpecId) {
+pub(super) fn smod(interpreter: &mut Interpreter, _host: &mut dyn Host) {
     gas!(interpreter, gas::LOW);
     pop_top!(interpreter, op1, op2);
     if *op2 != U256::ZERO {
@@ -49,21 +49,21 @@ pub(super) fn smod(interpreter: &mut Interpreter, _host: &mut dyn Host, _spec: S
     }
 }
 
-pub(super) fn addmod(interpreter: &mut Interpreter, _host: &mut dyn Host, _spec: SpecId) {
+pub(super) fn addmod(interpreter: &mut Interpreter, _host: &mut dyn Host) {
     gas!(interpreter, gas::MID);
     pop_top!(interpreter, op1, op2, op3);
     *op3 = op1.add_mod(op2, *op3)
 }
 
-pub(super) fn mulmod(interpreter: &mut Interpreter, _host: &mut dyn Host, _spec: SpecId) {
+pub(super) fn mulmod(interpreter: &mut Interpreter, _host: &mut dyn Host) {
     gas!(interpreter, gas::MID);
     pop_top!(interpreter, op1, op2, op3);
     *op3 = op1.mul_mod(op2, *op3)
 }
 
-pub(super) fn exp(interpreter: &mut Interpreter, _host: &mut dyn Host, spec: SpecId) {
+pub(super) fn exp<SPEC: Spec>(interpreter: &mut Interpreter, _host: &mut dyn Host) {
     pop_top!(interpreter, op1, op2);
-    gas_or_fail!(interpreter, gas::exp_cost(*op2, spec));
+    gas_or_fail!(interpreter, gas::exp_cost::<SPEC>(*op2));
     *op2 = op1.pow(*op2);
 }
 
@@ -82,7 +82,7 @@ pub(super) fn exp(interpreter: &mut Interpreter, _host: &mut dyn Host, spec: Spe
 /// `y | !mask` where `|` is the bitwise `OR` and `!` is bitwise negation. Similarly, if
 /// `b == 0` then the yellow paper says the output should start with all zeros, then end with
 /// bits from `b`; this is equal to `y & mask` where `&` is bitwise `AND`.
-pub(super) fn signextend(interpreter: &mut Interpreter, _host: &mut dyn Host, _spec: SpecId) {
+pub(super) fn signextend(interpreter: &mut Interpreter, _host: &mut dyn Host) {
     gas!(interpreter, gas::LOW);
     pop_top!(interpreter, op1, op2);
     if op1 < U256::from(32) {
diff --git a/crates/interpreter/src/instructions/bitwise.rs b/crates/interpreter/src/instructions/bitwise.rs
index 6e09945..700a38b 100644
--- a/crates/interpreter/src/instructions/bitwise.rs
+++ b/crates/interpreter/src/instructions/bitwise.rs
@@ -6,67 +6,67 @@ const fn btou256(b: bool) -> U256 {
     U256::from_limbs([b as u64, 0, 0, 0])
 }
 
-pub(super) fn lt(interpreter: &mut Interpreter, _host: &mut dyn Host, _spec: SpecId) {
+pub(super) fn lt(interpreter: &mut Interpreter, _host: &mut dyn Host) {
     gas!(interpreter, gas::VERYLOW);
     pop_top!(interpreter, op1, op2);
     *op2 = btou256(op1 < *op2);
 }
 
-pub(super) fn gt(interpreter: &mut Interpreter, _host: &mut dyn Host, _spec: SpecId) {
+pub(super) fn gt(interpreter: &mut Interpreter, _host: &mut dyn Host) {
     gas!(interpreter, gas::VERYLOW);
     pop_top!(interpreter, op1, op2);
     *op2 = btou256(op1 > *op2);
 }
 
-pub(super) fn slt(interpreter: &mut Interpreter, _host: &mut dyn Host, _spec: SpecId) {
+pub(super) fn slt(interpreter: &mut Interpreter, _host: &mut dyn Host) {
     gas!(interpreter, gas::VERYLOW);
     pop_top!(interpreter, op1, op2);
     *op2 = btou256(i256_cmp(op1, *op2) == Ordering::Less);
 }
 
-pub(super) fn sgt(interpreter: &mut Interpreter, _host: &mut dyn Host, _spec: SpecId) {
+pub(super) fn sgt(interpreter: &mut Interpreter, _host: &mut dyn Host) {
     gas!(interpreter, gas::VERYLOW);
     pop_top!(interpreter, op1, op2);
     *op2 = btou256(i256_cmp(op1, *op2) == Ordering::Greater);
 }
 
-pub(super) fn eq(interpreter: &mut Interpreter, _host: &mut dyn Host, _spec: SpecId) {
+pub(super) fn eq(interpreter: &mut Interpreter, _host: &mut dyn Host) {
     gas!(interpreter, gas::VERYLOW);
     pop_top!(interpreter, op1, op2);
     *op2 = btou256(op1 == *op2);
 }
 
-pub(super) fn iszero(interpreter: &mut Interpreter, _host: &mut dyn Host, _spec: SpecId) {
+pub(super) fn iszero(interpreter: &mut Interpreter, _host: &mut dyn Host) {
     gas!(interpreter, gas::VERYLOW);
     pop_top!(interpreter, op1);
     *op1 = btou256(*op1 == U256::ZERO);
 }
 
-pub(super) fn bitand(interpreter: &mut Interpreter, _host: &mut dyn Host, _spec: SpecId) {
+pub(super) fn bitand(interpreter: &mut Interpreter, _host: &mut dyn Host) {
     gas!(interpreter, gas::VERYLOW);
     pop_top!(interpreter, op1, op2);
     *op2 = op1 & *op2;
 }
 
-pub(super) fn bitor(interpreter: &mut Interpreter, _host: &mut dyn Host, _spec: SpecId) {
+pub(super) fn bitor(interpreter: &mut Interpreter, _host: &mut dyn Host) {
     gas!(interpreter, gas::VERYLOW);
     pop_top!(interpreter, op1, op2);
…